Move tagging read to queries/, rename use_cases/ to commands/
GetTagsForBookUseCase is the module's only read; it is a simple read
(ownership check + one find_by_book), so per ADR-0001's halfway option it
moves to queries/ unchanged, still delegating to the repository and
returning domain entities. The remaining package holds commands only and
renames to commands/ per the end-state naming plan; the
queries-are-dead-ends contract now covers both use_cases and commands
source wildcards.
