+++ backend/src/application/tagging/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging command use cases."""
