+++ backend/src/application/tagging/queries/__init__.py
@@ -1,0 +1,1 @@
+"""Tagging read use cases (query side)."""
