--- backend/src/application/tagging/use_cases/__init__.py
@@ -1,1 +1,0 @@
-"""Tagging use cases."""
